+++ packages/crayons-core/src/components/select-option/select-option.tsx
@@ -188,8 +188,8 @@ export class SelectOption {
         return (
           <Fragment>
             {checkbox}
-            {this.createIcon()}
-            {this.createConversationDescription()}
+            {this.createConversationIcon()}
+            {description}
             {selectedIconContainer}
           </Fragment>
         );
@@ -205,6 +205,24 @@ export class SelectOption {
   }
 
   createDescription() {
+    if (this.metaText) {
+      const metaTextDetails = [];
+      if (this.metaText?.name) metaTextDetails.push(this.metaText.name);
+      if (this.metaText?.email) metaTextDetails.push(this.metaText.email);
+      if (this.metaText?.mobile) metaTextDetails.push(this.metaText.mobile);
+
+      return (
+        <div class={'description ' + 'icon-margin '}>
+          <span class='description-text'>{this.text}</span>
+          {this.subText && (
+            <span class='description-subText-conversation'>{this.subText}</span>
+          )}
+          <span class='description-metaText-details'>
+            {metaTextDetails?.join(' | ')}
+          </span>
+        </div>
+      );
+    }
     return this.subText ? (
       <div
         class={
@@ -227,23 +245,8 @@ export class SelectOption {
     );
   }
 
-  createConversationDescription() {
-    const metaTextDetails = [];
-    if (this.metaText.name) metaTextDetails.push(this.metaText.name);
-    if (this.metaText.email) metaTextDetails.push(this.metaText.email);
-    if (this.metaText.mobile) metaTextDetails.push(this.metaText.mobile);
-
-    return this.subText ? (
-      <div class={'description ' + 'icon-margin '}>
-        <span class='description-text'>{this.text}</span>
-        <span class='description-subText-conversation'>{this.subText}</span>
-        <span class='description-metaText-details'>
-          {metaTextDetails?.join(' | ')}
-        </span>
-      </div>
-    ) : (
-      <span class={'description ' + 'icon-margin'}>{this.text}</span>
-    );
+  createConversationIcon() {
+    return <div class='conversation-icon'>{this.createIcon()}</div>;
   }
 
   createIcon() {